initial time 로직 추가

+++ app/src/main/java/com/example/pill_checker/data/Time.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Time(
-    @PrimaryKey val tid: Long,
+    @PrimaryKey val tid: Int,
     var count: Int
     )
